return S3 URI on success
